+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Bugfix: Call commands timing out now raise a timeout exception in the caller, but do not crash the actor
   * Bugfix: CallController#dial now unblocks immediately if the original call ends
   * Bugfix: CallController#dial now unblocks when the connected outbound call unjoins, rather than ending, incase post-processing on the outbound call is required
+  * Bugfix: CallController#dial now hangs up outbound legs when it unblocks
 
 # 2.0.0.alpha2 - 2012-01-30
   * Change: Plugins no longer load dialplan/event/rpc/console methods using corresponding class methods